requirements.txt
fastapi
uvicorn
motor
pydantic
passlib[bcrypt]
python-dotenv
python-jose[cryptography]
